refactor: api 분리 및 파일 구조 변경 2023.08.14 1. addProfile => uploadProfile 파일명 변경 2. api 폴더 추가 및 파일 분리
