+++ index.html
@@ -68,5 +68,6 @@
   </body>
   <script type="module" src="js/config/aws.js"></script>
   <script type="module" src="/js/constant.js"></script>
-  <script type="module" src="/js/addProfile.js"></script>
+  <script type="module" src="/js/main.js"></script>
+  <script type="module" src="/js/api/uploadProfile.js"></script>
 </html>
